simpleRag/pyproject.toml
[project]
name = "simplerag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.4",
    "dotenv>=0.9.9",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "ipynb>=0.5.1",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.0",
    "langchain-nomic>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.2",
    "markdown>=3.10",
    "pyjwt>=2.9.0",
    "unstructured>=0.18.18",
    "uvicorn>=0.34.0",
]
